+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.WellFed"
         tools:targetApi="31">
         <activity
-            android:name=".Recipe2.RecipeActivity"
+            android:name=".recipe.RecipeActivity"
             android:exported="false"
             android:parentActivityName=".MainActivity">
             <meta-data
