+++ app/src/main/java/com/example/wellfed/MainActivity.java
@@ -1,14 +1,14 @@
 package com.example.wellfed;
 
-import androidx.appcompat.app.AppCompatDelegate;
 import androidx.fragment.app.FragmentActivity;
 import androidx.viewpager2.widget.ViewPager2;
 
 import android.os.Bundle;
 import android.view.Menu;
 
-import com.example.wellfed.navigation.NavigationCollectionAdapter;
+
 import com.google.android.material.bottomappbar.BottomAppBar;
+import com.example.wellfed.navigation.NavigationCollectionAdapter;
 
 public class MainActivity extends FragmentActivity {
     final String TAG = "Sample";
