--- app/src/main/java/com/example/wellfed/Recipe2/Recipe.java
@@ -1,13 +1,0 @@
-package com.example.wellfed.Recipe2;
-
-// mock class for testing
-// Todo will require to be replaced
-public class Recipe {
-    private String description;
-    private String category;
-
-    public Recipe(String description, String category) {
-        this.description = description;
-        this.category = category;
-    }
-}
