--- app/src/main/java/com/example/wellfed/Recipe2/RecipeController.java
@@ -1,4 +1,0 @@
-package com.example.wellfed.Recipe2;
-
-public class RecipeController {
-}
