+++ app/src/main/java/com/example/wellfed/recipe/RecipeController.java
@@ -1,0 +1,4 @@
+package com.example.wellfed.recipe;
+
+public class RecipeController {
+}
