+++ app/src/main/java/com/example/wellfed/recipe/RecipeDB.java
@@ -1,0 +1,8 @@
+package com.example.wellfed.recipe;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class RecipeDB {
+
+}
